保全未完待续...

+++ NestSound/Src/NSHome/Controller/NSActivityJoinerListController.m
@@ -55,12 +55,12 @@
         self.page ++;
         self.requestParams = @{kIsLoadingMore:@(YES),
                                @"aid":self.aid,
-                               @"page":[NSString stringWithFormat:@"%ld",self.page]};
+                               @"page":[NSString stringWithFormat:@"%ld",(long)self.page]};
     }else{
         self.page = 1;
         self.requestParams = @{kIsLoadingMore:@(NO),
                                @"aid":self.aid,
-                               @"page":[NSString stringWithFormat:@"%ld",self.page]};
+                               @"page":[NSString stringWithFormat:@"%ld",(long)self.page]};
     }
     
     self.requestURL = joinedUserListUrl;
